Extract listening code to ApplicationRunner
